Introduce PatternGraphFactory Some code clean up
