+++ plugin/src/main/kotlin/org/jetbrains/research/plugin/PyFlowGraphBuildingUtils.kt
@@ -1,9 +1,8 @@
 package org.jetbrains.research.plugin
 
 import com.jetbrains.python.psi.PyFunction
-import org.jetbrains.research.plugin.jgrapht.createPatternSpecificGraph
+import org.jetbrains.research.plugin.jgrapht.PatternDirectedAcyclicGraph
 import org.jetbrains.research.plugin.jgrapht.edges.PatternSpecificMultipleEdge
-import org.jetbrains.research.plugin.jgrapht.loadPatternSpecificGraph
 import org.jetbrains.research.plugin.jgrapht.vertices.PatternSpecificVertex
 import org.jetbrains.research.plugin.pyflowgraph.GraphBuildingException
 import org.jetbrains.research.plugin.pyflowgraph.PyFlowGraphBuilder
@@ -17,11 +16,11 @@ fun buildPyFlowGraphForMethod(node: PyFunction, builder: String = "kotlin")
         "python" -> {
             val tempFile = createTempFileFromMethodPsi(node)
             val dotFile = buildPyFlowGraphBySubprocess(tempFile)
-            loadPatternSpecificGraph(dotFile.inputStream())
+            PatternDirectedAcyclicGraph(dotFile.inputStream())
         }
         "kotlin" -> {
             val methodPyFlowGraph = PyFlowGraphBuilder().buildForPyFunction(node)
-            createPatternSpecificGraph(methodPyFlowGraph)
+            PatternDirectedAcyclicGraph(methodPyFlowGraph)
         }
         else -> throw  IllegalArgumentException()
     }
